MAJ pour version Leed multi user

+++ fleaditlater/fleaditlater.plugin.disabled.php
@@ -10,7 +10,8 @@
 
 function fleaditlater_plugin_AddButton(&$event){
 	$eventId = $event->getId();
-	$count = mysql_query('SELECT COUNT(id) FROM '.MYSQL_PREFIX.'plugin_feaditlater WHERE event='.$eventId);
+	$myUser = (isset($_SESSION['currentUser'])?unserialize($_SESSION['currentUser']):false);
+	$count = mysql_query('SELECT COUNT(1) FROM '.MYSQL_PREFIX.'plugin_feaditlater WHERE event='.$eventId. ' and idUser='.$myUser->getId());
 	$count = mysql_fetch_row($count);
 	if(!$count[0]){
 		echo '<div  onclick="fleadItLater('.$eventId.',\'add\',this);" class="fleaditLaterButton">Lire + Tard</div>';
@@ -18,41 +19,45 @@ function fleaditlater_plugin_AddButton(&$event){
 }
 
 function fleaditlater_plugin_displayEvents(&$myUser){
-	$query = mysql_query('SELECT le.id,le.title,le.link FROM '.MYSQL_PREFIX.'event le INNER JOIN '.MYSQL_PREFIX.'plugin_feaditlater fil ON (le.id=fil.event)');
-	if($query!=null){
-	echo '<aside class="fleaditLaterMenu">
-				
-				<h3 class="left">A lire</h3>
-					<ul class="clear">  							  								  							  							  								  	
-					<li>
-						<ul> ';
-							
-							while($data = mysql_fetch_array($query)){
-							echo '<li> 
+	if($myUser!=false){
+		$query = mysql_query('SELECT le.id,le.title,le.link FROM '.$myUser->getPrefixDatabase().'event le INNER JOIN '.MYSQL_PREFIX.'plugin_feaditlater fil ON (le.id=fil.event) WHERE fil.idUser='.$myUser->getId());
+		if($query!=null){
+		echo '<aside class="fleaditLaterMenu">
+					
+					<h3 class="left">A lire</h3>
+						<ul class="clear">  							  								  							  							  								  	
+						<li>
+							<ul> ';
 								
-									<img src="plugins/fleaditlater/img/read_icon.png">
-						
-								<a title="'.$data['link'].'" href="'.$data['link'].'" target="_blank">
-									'.Functions::truncate($data['title'],40).'
-								</a>		  
-								<button class="right" onclick="fleadItLater('.$data['id'].',\'delete\',this)" style="margin-left:5px;">
-									<span title="marquer comme lu" alt="marquer comme lu">Lu</span>
-								</button>
-								</li>';
-							}
+								while($data = mysql_fetch_array($query)){
+								echo '<li> 
+									
+										<img src="plugins/fleaditlater/img/read_icon.png">
+							
+									<a title="'.$data['link'].'" href="'.$data['link'].'" target="_blank">
+										'.Functions::truncate($data['title'],40).'
+									</a>		  
+									<button class="right" onclick="fleadItLater('.$data['id'].',\'delete\',this)" style="margin-left:5px;">
+										<span title="marquer comme lu" alt="marquer comme lu">Lu</span>
+									</button>
+									</li>';
+								}
 
-						echo '</ul>
-						
-					</li>
-				</ul>
-			</aside>';
-			}
+							echo '</ul>
+							
+						</li>
+					</ul>
+				</aside>';
+				} else {
+					echo 'erreur plugin: Fleaditlater. Désactiver et réactiver le plugin';
+				}
+	}
 }
 
 function fleaditlater_plugin_action($_,$myUser){
 	if($myUser==false) exit('Vous devez vous connecter pour cette action.');
 	if($_['state']=='add'){
-		$return = mysql_query('INSERT INTO '.MYSQL_PREFIX.'plugin_feaditlater (event)VALUES(\''.$_['id'].'\')');
+		$return = mysql_query('INSERT INTO '.MYSQL_PREFIX.'plugin_feaditlater (event,idUser)VALUES(\''.$_['id'].'\',\''.$myUser->getId().'\')');
 	}else{
 		$return = mysql_query('DELETE FROM '.MYSQL_PREFIX.'plugin_feaditlater WHERE event=\''.$_['id'].'\'');
 	}
